# static/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API to interact with [Wakapi](https://wakapi.dev)\n\n## Authentication\nSet header `Authorization` to your API Key encoded as Base64 and prefixed with `Basic`\n**Example:** `Basic ODY2NDhkNzQtMTljNS00NTJiLWJhMDEtZmIzZWM3MGQ0YzJmCg==`",
        "title": "Wakapi API",
        "contact": {
            "name": "Ferdinand Mütsch",
            "url": "https://github.com/muety",
            "email": "ferdinand@muetsch.io"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://github.com/muety/wakapi/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/relay": {
            "get": {
                "tags": [
                    "relay"
                ],
                "summary": "Proxy an GET API request to another Wakapi instance",
                "operationId": "relay-get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Original URL to perform the request to",
                        "name": "X-Target-URL",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Returned if request path is not whitelisted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Returned if upstream host is down",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "relay"
                ],
                "summary": "Proxy an PUT API request to another Wakapi instance",
                "operationId": "relay-put",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Original URL to perform the request to",
                        "name": "X-Target-URL",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Returned if request path is not whitelisted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Returned if upstream host is down",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "relay"
                ],
                "summary": "Proxy an POST API request to another Wakapi instance",
                "operationId": "relay-post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Original URL to perform the request to",
                        "name": "X-Target-URL",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Returned if request path is not whitelisted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Returned if upstream host is down",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "relay"
                ],
                "summary": "Proxy an DELETE API request to another Wakapi instance",
                "operationId": "relay-delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Original URL to perform the request to",
                        "name": "X-Target-URL",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Returned if request path is not whitelisted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Returned if upstream host is down",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "relay"
                ],
                "summary": "Proxy an PATCH API request to another Wakapi instance",
                "operationId": "relay-patch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Original URL to perform the request to",
                        "name": "X-Target-URL",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Returned if request path is not whitelisted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Returned if upstream host is down",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}